Direct3D11Matrix が内部的に XMMATRIX を使うように修正。 blue-sky レイヤーが XMMATRIX, XMVECTOR を直接使わず、 Direct3D11Matrix, Direct3D11Vector を使うように修正。 主なコンパイラの警告を修正。
Direct3D11Matrix が内部で保持している XMMATRIX は 16 バイト境界にアライメントする必要がある。そのため Direct3D11Matrix の new, delete 演算子をオーバーロードしている。
また、 SkinningAnimationSet::MatrixList を Direct3D11Matrix　の vector から Direct3D11Matrix* の　vector に変更。

+++ source/OggVorbisFile.cpp
@@ -21,10 +21,10 @@ OggVorbisFile::OggVorbisFile( const char* file_name )
 	vorbis_info_ = ov_info( & file_, -1 );
 
 	format_.wFormatTag = WAVE_FORMAT_PCM;
-	format_.nChannels = vorbis_info_->channels;
+	format_.nChannels = static_cast< WORD >( vorbis_info_->channels );
 	format_.nSamplesPerSec = vorbis_info_->rate;
     format_.nAvgBytesPerSec = 2 * vorbis_info_->rate * vorbis_info_->channels;
-    format_.nBlockAlign = 2 * vorbis_info_->channels;
+    format_.nBlockAlign = static_cast< WORD >( 2 * vorbis_info_->channels );
     format_.wBitsPerSample = 16;
     format_.cbSize = 0;
 }
